requirements.txt
# Core Django
Django==5.2.4
djangorestframework==3.16.0

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# API Documentation
drf-yasg==1.21.10

# CORS Support
django-cors-headers==4.7.0

# Environment Variables
python-dotenv==1.1.1

# Database
psycopg2-binary==2.9.11

# Image Processing
Pillow==11.0.0

# Production Server
whitenoise==6.9.0
gunicorn==23.0.0
